feat: integrate Together AI for weather-based suggestions
Integrates an AI-powered recommendation feature using the Together AI API.

This commit introduces the following changes:
- Adds a new `AiSuggestionRepository` to fetch clothing and safety advice based on weather data.
- Implements `TogetherApi` using Retrofit to communicate with the Together AI service.
- Creates a `WeatherPromptBuilder` to construct a detailed prompt from current and hourly weather data.
- Adds necessary data models for the AI API request and response.
- Includes hourly forecast data in the OpenWeatherMap API call to enrich the AI prompt.
- Updates the ViewModel and MainActivity to process and display the AI-generated suggestions.
- Adds the Together AI API key to the build configuration.

+++ app/src/main/java/com/example/androidweatherapp/model/WeatherData.kt
@@ -100,7 +100,7 @@ data class WeatherResponse(
         val pop: Double,
         val pressure: Int,
         val temp: Double,
-        val uvi: Int,
+        val uvi: Double,
         val visibility: Int,
         val weather: List<Weather>,
         val wind_deg: Int,
